requirements.txt
tabulate
click
matplotlib
pyyaml
pandas
wheel
setuptools
numpy
